LayoutTests/platform/ios-sim-deprecated/http/tests/security/dataURL/xss-DENIED-from-data-url-to-data-url-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "null" from accessing a frame with origin "null". Protocols, domains, and ports must match.
This tests that a data: URL subframe can't access a child data: URL subframe of itself.



--------
Frame: 'aFrame'
--------
Inner iframe.

Test


--------
Frame: 'inner'
--------
PASS: Cross frame access from a data: URL was denied.

Inner-inner iframe.
